requirements.txt
Flask==2.0.2
matplotlib==3.4.3
nltk==3.8.1
numpy==1.21.2
pandas==1.3.4
python_dateutil==2.8.2
scikit_learn==1.2.2
spacy==3.5.3
xgboost==1.7.5
